Move `instances/{instance_lock,lock_file}.{cpp,h}
... to a new subdirectory.

This is one step to making the filesystem organization match the build
system.

Bug: b/392949844
Test: bazel test '//...'
