+++ base/cvd/cuttlefish/host/commands/cvd/acloud/BUILD.bazel
@@ -38,8 +38,8 @@ cc_library(
         "//cuttlefish/host/commands/cvd/cli/selector:parser",
         "//cuttlefish/host/commands/cvd/cli:types",
         "//cuttlefish/host/commands/cvd/cli:utils",
+        "//cuttlefish/host/commands/cvd/instances/lock",
         "//cuttlefish/host/commands/cvd/instances:instance_database_utils",
-        "//cuttlefish/host/commands/cvd/instances:lock",
         "//cuttlefish/host/commands/cvd/utils",
         "//cuttlefish/host/libs/config",
         "//libbase",
