+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/BUILD.bazel
@@ -25,9 +25,9 @@ cc_library(
         "//cuttlefish/common/libs/utils",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/commands/cvd/cli:types",
+        "//cuttlefish/host/commands/cvd/instances/lock",
         "//cuttlefish/host/commands/cvd/instances:cvd_persistent_data",
         "//cuttlefish/host/commands/cvd/instances:instance_database_utils",
-        "//cuttlefish/host/commands/cvd/instances:lock",
         "//cuttlefish/host/commands/cvd/utils",
         "//cuttlefish/host/libs/config",
         "//libbase",
