+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/creation_analyzer.h
@@ -26,7 +26,7 @@
 #include "common/libs/utils/result.h"
 #include "common/libs/utils/unique_resource_allocator.h"
 #include "cuttlefish/host/commands/cvd/instances/cvd_persistent_data.pb.h"
-#include "host/commands/cvd/instances/instance_lock.h"
+#include "host/commands/cvd/instances/lock/instance_lock.h"
 #include "host/commands/cvd/cli/selector/start_selector_parser.h"
 
 namespace cuttlefish {
