+++ base/cvd/cuttlefish/host/commands/cvd/instances/BUILD.bazel
@@ -13,29 +13,6 @@ cc_proto_library(
 )
 
 cc_library(
-    name = "lock",
-    srcs = [
-        "instance_lock.cpp",
-        "lock_file.cpp",
-    ],
-    hdrs = [
-        "instance_lock.h",
-        "lock_file.h",
-    ],
-    copts = [
-        "-Werror=sign-compare",
-        "-std=c++17",
-    ],
-    strip_include_prefix = "//cuttlefish",
-    deps = [
-        "//cuttlefish/common/libs/fs",
-        "//cuttlefish/common/libs/utils",
-        "//cuttlefish/common/libs/utils:result",
-        "//libbase",
-    ],
-)
-
-cc_library(
     name = "instances",
     srcs = [
         "data_viewer.cpp",
@@ -63,7 +40,6 @@ cc_library(
     deps = [
         ":cvd_persistent_data",
         ":instance_database_utils",
-        ":lock",
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils",
         "//cuttlefish/common/libs/utils:result",
@@ -72,6 +48,7 @@ cc_library(
         "//cuttlefish/host/commands/cvd/cli:command_request",
         "//cuttlefish/host/commands/cvd/cli:types",
         "//cuttlefish/host/commands/cvd/cli:utils",
+        "//cuttlefish/host/commands/cvd/instances/lock",
         "//cuttlefish/host/commands/cvd/legacy:cvd_server_cc_proto",
         "//cuttlefish/host/commands/cvd/utils",
         "//cuttlefish/host/libs/config",
