+++ base/cvd/cuttlefish/host/commands/cvd/instances/lock/BUILD.bazel
@@ -1,0 +1,26 @@
+package(
+    default_visibility = ["//:android_cuttlefish"],
+)
+
+cc_library(
+    name = "lock",
+    srcs = [
+        "instance_lock.cpp",
+        "lock_file.cpp",
+    ],
+    hdrs = [
+        "instance_lock.h",
+        "lock_file.h",
+    ],
+    copts = [
+        "-Werror=sign-compare",
+        "-std=c++17",
+    ],
+    strip_include_prefix = "//cuttlefish",
+    deps = [
+        "//cuttlefish/common/libs/fs",
+        "//cuttlefish/common/libs/utils",
+        "//cuttlefish/common/libs/utils:result",
+        "//libbase",
+    ],
+)
